move postgres code to root, remove mongo
