--- src/db/mongoose.js
@@ -1,9 +1,0 @@
-const mongoose = require('mongoose');
-
-if (process.env.NODE_ENV !== 'test') {
-  mongoose.connect(process.env.MONGODB_URI, { useNewUrlParser: true });
-}
-
-module.exports = {
-  mongoose
-};
