+++ src/worker/updateAllTimeStats.js
@@ -1,46 +1,15 @@
-const AppYear = require('../models/AppYear');
-
-exports.updateAllTimeStats = async () => {
+module.exports = async ({ sequelize }) => {  
   console.log('updating all time stats');
-
-  await AppYear.aggregate([
-    {
-      $group: {
-        _id: '$appId',
-        years: {
-          $push: '$$ROOT'
-        }
-      }
-    },
-    {
-      $addFields: {
-        average: {
-          $avg: {
-            $map: {
-              input: '$years',
-              in: '$$this.average'
-            }
-          }
-        },
-        peak: {
-          $max: {
-            $map: {
-              input: '$years',
-              in: '$$this.peak'
-            }
-          }
-        }
-      }
-    },
-    {
-      $project: {
-        years: 0
-      }
-    },
-    {
-      $merge: { into: 'apps' }
-    }
-  ]).allowDiskUse(true);
-
-  console.log('finished updating all time stats');
-};
+  await sequelize.query(`
+    UPDATE apps
+    SET (average, peak) = (
+      SELECT
+        AVG(average) AS average,
+        MAX(peak) AS peak
+      FROM app_years
+      WHERE app_years.app_id=apps.id
+      GROUP BY app_years.app_id
+    )
+  `);
+  console.log('all time stats successfully updated');
+}
